test: split coordinator tests by behavior
